% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinityPropagation.R
\name{affinityPropagation}
\alias{affinityPropagation}
\title{affinityPropagation}
\usage{
affinityPropagation(ap.data, sim = "Jaccard")
}
\arguments{
\item{sim}{method for compute similarity between gene set:
default is 'Jaccard'}

\item{apdata:}{a list of input data
$ output.dir
$ genesetInfo
$ genesetIds}
}
\description{
affinityPropagation
}
