% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumer.R
\name{sumer}
\alias{sumer}
\title{summarizing multiomics enrichment analysis results}
\usage{
sumer(config_file, output_dir, n_threads = 4)
}
\arguments{
\item{config_file:}{specifiy input data locations}

\item{output_dir:}{output directory}
}
\description{
summarizing multiomics enrichment analysis results
}
